+++ recipe-front-end/src/components/menu-planner/DayDetails.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { connect } from 'react-redux';
 import { Localisation } from "../../localisation/AppTexts";
 import { DayMenu, ReduxModel } from "../../redux/Store";
-import DroppableMenuItem from "./DroppableIngredient";
+import DroppableMenuItem from "./DroppableMenuItem";
 import { filterForDate } from "./MenuPlanner";
 
 interface OwnProps {
