+++ recipe-front-end/src/components/menu-planner/DroppableMenuItem.tsx
